Added development mode and fixed missing libraries

+++ events/onLeaveRemoveEntry.evt.js
@@ -11,10 +11,14 @@ const { MongoClient } = require("mongodb");
  * @param {*} RM
  */
 async function runEvent(RM, ...args) {
+  if (args[0].user.bot) return;
+  if (args[0].guild.id !== global.app.config.mainGuild) return;
   const client = new MongoClient(global.mongoConnectionString);
   try {
     const database = client.db("IRIS");
-    const userdata = database.collection("userdata");
+    const userdata = database.collection(
+      global.app.config.development ? "userdata_dev" : "userdata"
+    );
     const result = await userdata.deleteOne({ id: args[0].id });
     /* prettier-ignore */
     global.app.debugLog(chalk.white.bold("["+moment().format("M/D/y HH:mm:ss")+"] ["+returnFileName()+"] ")+ global.chalk.yellow(args[0].user.tag) + " has left the server." + (result.deletedCount>0 ? " Their entry has been removed from the database.":""))
